+++ lib/l10n/app_localizations_en.dart
@@ -571,8 +571,20 @@ class AppLocalizationsEn extends AppLocalizations {
   String get cancelTransaction => 'Cancel Transaction';
 
   @override
-  String get speedUpTip =>
-      'Normally, a transaction takes <light>3 minutes</light> to confirm. However, when the network is congested and your transaction may take longer than 3 minutes or more, you can speed up the transaction process by increasing the transaction fee.';
+  String speedUpTip(String slotTime) {
+    return 'Normally, a transaction takes <light>$slotTime</light> to confirm. When the network is congested, your transaction may take $slotTime or longer. You can speed up the transaction process by increasing the transaction fee.';
+  }
+
+  @override
+  String nMinutes(num count, String formattedMinutes) {
+    String _temp0 = intl.Intl.pluralLogic(
+      count,
+      locale: localeName,
+      other: '$formattedMinutes minutes',
+      one: '$formattedMinutes minute',
+    );
+    return '$_temp0';
+  }
 
   @override
   String get transactionCancelTip =>